apps/backend/requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
django-filter>=24.0,<25.0
django-celery-results>=2.5,<3.0
celery>=5.4,<6.0
redis>=5.0,<6.0
psycopg[binary]>=3.2,<4.0
gunicorn>=22.0,<23.0
pgvector>=0.3,<1.0
stripe>=10.0,<12.0
PyGithub>=2.3.0
cryptography>=42.0.0
